requirements.txt
numpy>=1.15.4
matplotlib>=3.0.2
tensorflow-gpu>=1.12
